+++ src/main/java/cafe/Factory.java
@@ -2,8 +2,8 @@ package cafe;
 
 import cafe.database.ConnectionPool;
 import cafe.database.RealJdbcConnectionPool;
-import cafe.questions.ArticleRepository;
-import cafe.questions.MemoryArticleRepository;
+import cafe.questions.repository.ArticleRepository;
+import cafe.questions.repository.MemoryArticleRepository;
 import cafe.users.repository.JdbcUserRepository;
 import cafe.users.repository.UserRepository;
 import cafe.util.PropertiesLoader;
